+++ e2e/e2e-app/mitosis.config.cjs
@@ -1,5 +1,5 @@
 /**
- * @type {import('@builder.io/mitosis'.MitosisConfig)}
+ * @type {import('@builder.io/mitosis').MitosisConfig}
  */
 module.exports = {
   files: 'src/**',
@@ -45,21 +45,7 @@ module.exports = {
   },
   options: {
     angular: {
-      preserveImports: true,
-      importMapper: (component, theImport, importedValues, componentsUsed) => {
-        if (theImport.path.endsWith('.lite')) {
-          const cleanPath = theImport.path.replaceAll('-', '').replace('.lite', '').toLowerCase();
-
-          const component = componentsUsed.find((componentUsed) => {
-            return cleanPath.includes(componentUsed.toLowerCase());
-          });
-          if (component) {
-            return `import {${component}Module} from "${theImport.path.replace('.lite', '')}";`;
-          }
-        }
-
-        return `import ${importedValues.namedImports} from '${theImport.path}';`;
-      },
+      api: 'signals',
     },
     react: { transpiler: { format: 'esm', languages: ['ts'] } },
     stencil: {