+++ packages/core/src/__tests__/builder/builder.test.ts
@@ -15,23 +15,24 @@ import { compileAwayBuilderComponents } from '@/plugins/compile-away-builder-com
 import { BuilderContent } from '@builder.io/sdk';
 
 import { componentToAngular } from '@/generators/angular';
-import columns from './data/blocks/columns.raw.tsx?raw';
-import customCode from './data/blocks/custom-code.raw.tsx?raw';
-import embed from './data/blocks/embed.raw.tsx?raw';
-import image from './data/blocks/image.raw.tsx?raw';
-import indexInFor from './data/blocks/index-in-for.raw.tsx?raw';
-import stamped from './data/blocks/stamped-io.raw.tsx?raw';
-import booleanContent from './data/builder/boolean.json?raw';
-import customComponentSlotPropertyContent from './data/builder/custom-component-slot-property.json?raw';
-import customComponentTags from './data/builder/custom-component-tags.json?raw';
-import lazyLoadSection from './data/builder/lazy-load-section.json?raw';
-import localization from './data/builder/localization.json?raw';
-import slotsContent from './data/builder/slots.json?raw';
-import slots2Content from './data/builder/slots2.json?raw';
-import textBindings from './data/builder/text-bindings.json?raw';
-import advancedFor from './data/for/advanced-for.raw.tsx?raw';
-import asyncBindings from './data/ref/basic-ref-assignment.raw.tsx?raw';
-import show from './data/show/show-expressions.raw.tsx?raw';
+import columns from '../data/blocks/columns.raw.tsx?raw';
+import customCode from '../data/blocks/custom-code.raw.tsx?raw';
+import embed from '../data/blocks/embed.raw.tsx?raw';
+import image from '../data/blocks/image.raw.tsx?raw';
+import indexInFor from '../data/blocks/index-in-for.raw.tsx?raw';
+import stamped from '../data/blocks/stamped-io.raw.tsx?raw';
+import booleanContent from '../data/builder/boolean.json?raw';
+import customComponentSlotPropertyContent from '../data/builder/custom-component-slot-property.json?raw';
+import customComponentTags from '../data/builder/custom-component-tags.json?raw';
+import lazyLoadSection from '../data/builder/lazy-load-section.json?raw';
+import localization from '../data/builder/localization.json?raw';
+import slotsContent from '../data/builder/slots.json?raw';
+import slots2Content from '../data/builder/slots2.json?raw';
+import tagNameContent from '../data/builder/tag-name.json?raw';
+import textBindings from '../data/builder/text-bindings.json?raw';
+import advancedFor from '../data/for/advanced-for.raw.tsx?raw';
+import asyncBindings from '../data/ref/basic-ref-assignment.raw.tsx?raw';
+import show from '../data/show/show-expressions.raw.tsx?raw';
 
 const mitosisOptions: ToMitosisOptions = {
   format: 'legacy',
@@ -237,6 +238,26 @@ describe('Builder', () => {
       component: backToMitosis,
     });
     expect(mitosis.trim()).toEqual(code.trim());
+  });
+
+  test('Fragment', () => {
+    const code = dedent`
+      import { useStore } from "@builder.io/mitosis";
+
+      export default function MyComponent(props) {
+        const state = useStore({ people: ["Steve", "Sewell"] });
+
+        return (
+          <Fragment>
+            Hello!
+          </Fragment>
+        );
+      }
+    `;
+
+    const component = parseJsx(code);
+    const builderJson = componentToBuilder()({ component });
+    expect(builderJson).toMatchSnapshot();
   });
 
   test('Regenerate loop', () => {
@@ -509,6 +530,132 @@ describe('Builder', () => {
 
     expect(component).toMatchSnapshot();
     expect(out).toMatchSnapshot();
+  });
+
+  test('tagName', async () => {
+    const component = builderContentToMitosisComponent(JSON.parse(tagNameContent));
+    expect(component).toMatchInlineSnapshot(`
+      {
+        "@type": "@builder.io/mitosis/component",
+        "children": [
+          {
+            "@type": "@builder.io/mitosis/node",
+            "bindings": {
+              "num": {
+                "bindingType": "expression",
+                "code": "10",
+                "type": "single",
+              },
+            },
+            "children": [
+              {
+                "@type": "@builder.io/mitosis/node",
+                "bindings": {},
+                "children": [],
+                "meta": {},
+                "name": "CounterComponent",
+                "properties": {
+                  "$tagName": "counter",
+                },
+                "scope": {},
+                "slots": {},
+              },
+            ],
+            "meta": {},
+            "name": "ProgressBar",
+            "properties": {
+              "$name": "ProgressBar",
+              "$tagName": "progress-bar",
+            },
+            "scope": {},
+            "slots": {},
+          },
+        ],
+        "context": {
+          "get": {},
+          "set": {},
+        },
+        "exports": {},
+        "hooks": {
+          "onEvent": [],
+          "onMount": [],
+        },
+        "imports": [],
+        "inputs": undefined,
+        "meta": {
+          "useMetadata": {
+            "httpRequests": undefined,
+          },
+        },
+        "name": "MyComponent",
+        "refs": {},
+        "state": {},
+        "subComponents": [],
+      }
+    `);
+
+    const react = await componentToReact({})({ component });
+    const vue = await componentToVue({})({ component });
+    const angular = await componentToAngular({})({ component });
+
+    expect(react).toMatchInlineSnapshot(`
+      "import * as React from \\"react\\";
+
+      function MyComponent(props) {
+        return (
+          <ProgressBar num={10}>
+            <CounterComponent />
+          </ProgressBar>
+        );
+      }
+
+      export default MyComponent;
+      "
+    `);
+
+    expect(vue).toMatchInlineSnapshot(`
+      "<template>
+        <progress-bar :num=\\"10\\"><counter></counter></progress-bar>
+      </template>
+
+      <script>
+      import { defineComponent } from \\"vue\\";
+
+      export default defineComponent({
+        name: \\"my-component\\",
+      });
+      </script>"
+    `);
+
+    expect(angular).toMatchInlineSnapshot(`
+      "import { NgModule } from \\"@angular/core\\";
+      import { CommonModule } from \\"@angular/common\\";
+
+      import { Component } from \\"@angular/core\\";
+
+      @Component({
+        selector: \\"my-component\\",
+        template: \`
+          <progress-bar [num]=\\"10\\"><counter></counter></progress-bar>
+        \`,
+        styles: [
+          \`
+            :host {
+              display: contents;
+            }
+          \`,
+        ],
+      })
+      export default class MyComponent {}
+
+      @NgModule({
+        declarations: [MyComponent],
+        imports: [CommonModule, ProgressBarModule, CounterComponentModule],
+        exports: [MyComponent],
+      })
+      export class MyComponentModule {}
+      "
+    `);
   });
 
   test('bindings', () => {
@@ -1325,8 +1472,7 @@ describe('Builder', () => {
                         "bindings": {},
                       },
                       "component": {
-                        "name": "Fragment",
-                        "options": {},
+                        "name": "Core:Fragment",
                       },
                     },
                   ],