+++ packages/core/src/generators/mitosis/generator.ts
@@ -21,10 +21,8 @@ import {
 import { MitosisComponent } from '@/types/mitosis-component';
 import { MitosisNode, checkIsForNode, checkIsShowNode } from '@/types/mitosis-node';
 import { TranspilerGenerator } from '@/types/transpiler';
-import * as babel from '@babel/core';
-import tsPlugin from '@babel/plugin-syntax-typescript';
 import json5 from 'json5';
-import { format } from 'prettier/standalone';
+import { check, format } from 'prettier/standalone';
 
 import { blockToReact, componentToReact } from '../react';
 
@@ -128,8 +126,17 @@ export const blockToMitosis = (
     if (text.includes('<') || text.includes('>')) {
       // test if this can parse as jsx
       try {
-        babel.parse(`let _ = <>${text}</>;`, {
-          plugins: [[tsPlugin, { isTSX: true }]],
+        /**
+         * We intentionally use the typescript parser here because texts like ">" will crash
+         * in the typescript parser but will not crash in the babel parser. The Prettier
+         * formatting that is run after JSX is generated also uses the typescript parser,
+         * so we want to make sure that doesn't crash.
+         */
+        check(`let _ = <>${text}</>;`, {
+          parser: 'typescript',
+          plugins: [
+            require('prettier/parser-typescript'), // To support running in browsers
+          ],
         });
         isInvalidJsx = false;
       } catch (e) {