+++ packages/core/src/generators/qwik/index.ts
@@ -2,3 +2,4 @@ export { addCommonStyles, addComponent, createFileSet } from './component';
 export type { FileSet, QwikOptions } from './component';
 export { componentToQwik } from './component-generator';
 export { File } from './src-generator';
+export * from './types';