+++ packages/core/src/generators/svelte/index.ts
@@ -1,1 +1,2 @@
 export { componentToSvelte } from './svelte';
+export * from './types';