+++ packages/core/src/helpers/babel-transform.ts
@@ -53,7 +53,7 @@ const handleErrorOrExpression = <VisitorContextType = any>({
 
     return result;
   } catch (err) {
-    // console.error('Error parsing code:\n', { code, result, useCode });
+    console.error('Error parsing code:\n', { code, result, useCode });
     throw err;
   }
 };